rename test folder to integration
