+++ integration/main_test.go
@@ -1,4 +1,4 @@
-package test
+package integration
 
 import (
 	"context"
